Introduce a makefile and split source into methods
Contains all targets necessary to get started building the test program.
Original header of source file for encryption have been split into their respective methods to illustrate combining multiple object files into a static library.
Holes are left in the makefile for the students to fill: targets for the rest of the object files for the complete encryption library and targets to build the command line application.