Fix TEST_STR invoking function twice.
Classic mistake of a macro using a parameters more than once and if it is a function, having the function invoked more than once.
Inserted a temporary variable to avoid the issue. Not the best resolution.
Also had to adjust the digest of test/main.c in the correction script.